requirements.txt
Cheetah3==3.2.5
CherryPy==18.6.0
